refactor: 合并动画条件相关定义，优化动画状态机和行为接口结构
